backups/20250127_160003/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
redis>=4.5.0

# AI Services
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
openai-whisper>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
librosa>=0.10.0

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
loguru>=0.7.0

# Web framework
websockets>=11.0.0
python-multipart>=0.0.6

# Database
supabase>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
asyncio>=3.4.3
webrtcvad>=2.0.10 